don't kill dynamodb local if CIRCLECI is set

+++ run-local.sh
@@ -2,9 +2,9 @@
 echo "killing dynamo if already running"
 pkill -f DynamoDBLocal
 
-echo "starting dynamo"
-./web-api/start-dynamo.sh &
-DYNAMO_PID=$!
+  echo "starting dynamo"
+  ./web-api/start-dynamo.sh &
+  DYNAMO_PID=$!
 
 node ./web-api/start-s3rver &
 S3RVER_PID=$!
@@ -50,5 +50,7 @@ npx sls offline start "$@" --config web-api/serverless-trial-sessions.yml &
 echo "starting proxy"
 node ./web-api/proxy.js
 
-pkill -P $DYNAMO_PID
+if [ ! -e $CIRCLECI ]; then 
+  pkill -P $DYNAMO_PID
+fi 
 kill $S3RVER_PID
